% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddVolumes.R
\name{AddVolumes}
\alias{AddVolumes}
\title{AddVolumes}
\usage{
AddVolumes(raw_data, start_well, stop_well, volume_added, sample_volume)
}
\arguments{
\item{raw_data}{A \code{data.table} of raw Clariostar readings.}

\item{start_well}{The first well in the series}

\item{stop_well}{The last well in the series}

\item{volume_added}{The volume of sample added to the Qubit reaction for
wells in this series}

\item{sample_volume}{The original sample volume for wells in this series (for
amount calculations)}
}
\value{
\strong{Modifies the \code{raw_data} \code{data.table}} by adding
  \code{volume_added} and \code{sample_volume} to the specified wells.
}
\description{
Add the volume added to the Qubit reaction and the original sample volume to
the raw data. Assumes plates were used in a "vertical" layout, \emph{e.g.}
with an 8-channel pipette.
}
