% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateAmounts.R
\name{CalculateAmounts}
\alias{CalculateAmounts}
\title{CalculateAmounts}
\usage{
CalculateAmounts(background_subtracted, standard_wells)
}
\arguments{
\item{background_subtracted}{A \code{data.table} of background-subtracted
Clariostar readings. Must have a column called
\code{background_subtracted}.}

\item{standard_wells}{A \code{data.table} of standards with the columns
\code{well} and \code{amount}, \emph{e.g.}
\preformatted{data.table(well = paste0(LETTERS[3:8], 5),
           amount = c(rep(10, 3), rep(100, 3)))}
    \tabular{lrr}{ \tab well \tab amount \cr
              1: \tab C5 \tab 10 \cr
              2: \tab D5 \tab 10 \cr
              3: \tab E5 \tab 10 \cr
              4: \tab F5 \tab 100 \cr
              5: \tab G5 \tab 100 \cr
              6: \tab H5 \tab 100 \cr}}
}
\value{
Returns a \code{data.table} of with an additional column called
  \code{calculated_amount} containing the result of \code{predict} on the
  \code{background_subtracted} values using the \code{lm} from the standards
}
\description{
Use the \code{background_subtracted} fluorescence values to make a \code{lm}
from the \code{standard_wells} and \code{predict} the amounts in the other
wells.
}
