% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractResults.R
\name{ExtractResults}
\alias{ExtractResults}
\title{ExtractResults}
\usage{
ExtractResults(calculated_amounts)
}
\arguments{
\item{calculated_amounts}{A \code{data.table} output from
\code{\link{CalculateAmounts}}}
}
\value{
A \code{data.table} with calculated_concentration and sample_volume
}
\description{
Convenience function for calculating the original concentration and total
amount from the results from \code{\link{CalculateAmounts}}
}
