% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitWellName.R
\name{SplitWellName}
\alias{SplitWellName}
\title{SplitWellName}
\usage{
SplitWellName(well_name)
}
\arguments{
\item{well_name}{A well name, \emph{e.g.} "A1", "B12"}
}
\value{
Returns a `data.table` of `plate_row` and `plate_col`
}
\description{
Split a well name into rows and columns
}
